fix: appstate isolation (prevent styles bleeding), loader no shrink
