+++ src/modules/ui/dom.ts
@@ -1,7 +1,7 @@
 import * as Utils from "@utils";
 import * as Constants from "@constants";
 import * as Observer from "./observer";
-import { AppState } from "@/index";
+import { AppState } from "@core/appState";
 import { animEngineState, getResumeScrollElement, reflow, toMs } from "@modules/ui/animationEngine";
 
 type SyncType = "syllable" | "word" | "line" | "unsynced";
@@ -11,6 +11,13 @@ const syncTypeIcons: Record<SyncType, string> = {
   word: `<svg width="14" height="14" viewBox="0 0 1024 1024" fill="currentColor" xmlns="http://www.w3.org/2000/svg"><rect x="636" y="239" width="389.981" height="233.271" rx="48" fill-opacity="0.5"/><path d="M0 335C0 289.745 0 267.118 14.0589 253.059C28.1177 239 50.7452 239 96 239H213C243.17 239 258.255 239 267.627 248.373C277 257.745 277 272.83 277 303V408C277 438.17 277 453.255 267.627 462.627C258.255 472 243.17 472 213 472H96C50.7452 472 28.1177 472 14.0589 457.941C0 443.882 0 421.255 0 376V335Z"/><path d="M337 304C337 273.83 337 258.745 346.373 249.373C355.745 240 370.83 240 401 240H460C505.255 240 527.882 240 541.941 254.059C556 268.118 556 290.745 556 336V377C556 422.255 556 444.882 541.941 458.941C527.882 473 505.255 473 460 473H401C370.83 473 355.745 473 346.373 463.627C337 454.255 337 439.17 337 409V304Z"/><rect y="552.271" width="1024" height="233" rx="48" fill-opacity="0.5"/></svg>`,
   line: `<svg width="14" height="14" viewBox="0 0 1024 1024" fill="currentColor" xmlns="http://www.w3.org/2000/svg"><rect x="636" y="239" width="389.981" height="233.271" rx="48"/><path d="M0 335C0 289.745 0 267.118 14.0589 253.059C28.1177 239 50.7452 239 96 239H213C243.17 239 258.255 239 267.627 248.373C277 257.745 277 272.83 277 303V408C277 438.17 277 453.255 267.627 462.627C258.255 472 243.17 472 213 472H96C50.7452 472 28.1177 472 14.0589 457.941C0 443.882 0 421.255 0 376V335Z"/><path d="M337 304C337 273.83 337 258.745 346.373 249.373C355.745 240 370.83 240 401 240H460C505.255 240 527.882 240 541.941 254.059C556 268.118 556 290.745 556 336V377C556 422.255 556 444.882 541.941 458.941C527.882 473 505.255 473 460 473H401C370.83 473 355.745 473 346.373 463.627C337 454.255 337 439.17 337 409V304Z"/><rect y="552.271" width="1024" height="233" rx="48" fill-opacity="0.5"/></svg>`,
   unsynced: `<svg width="14" height="14" viewBox="0 0 1024 1024" fill="currentColor" xmlns="http://www.w3.org/2000/svg"><rect x="636" y="239" width="389.981" height="233.271" rx="48" fill-opacity="0.5"/><path d="M0 335C0 289.745 0 267.118 14.0589 253.059C28.1177 239 50.7452 239 96 239H213C243.17 239 258.255 239 267.627 248.373C277 257.745 277 272.83 277 303V408C277 438.17 277 453.255 267.627 462.627C258.255 472 243.17 472 213 472H96C50.7452 472 28.1177 472 14.0589 457.941C0 443.882 0 421.255 0 376V335Z" fill-opacity="0.5"/><path d="M337 304C337 273.83 337 258.745 346.373 249.373C355.745 240 370.83 240 401 240H460C505.255 240 527.882 240 541.941 254.059C556 268.118 556 290.745 556 336V377C556 422.255 556 444.882 541.941 458.941C527.882 473 505.255 473 460 473H401C370.83 473 355.745 473 346.373 463.627C337 454.255 337 439.17 337 409V304Z" fill-opacity="0.5"/><rect y="552.271" width="1024" height="233" rx="48" fill-opacity="0.5"/></svg>`,
+};
+
+const syncTypeColors: Record<SyncType, string> = {
+  syllable: "#fde69b",
+  word: "#aad1ff",
+  line: "#c9f8da",
+  unsynced: "rgba(255, 255, 255, 0.7)",
 };
 
 function parseSvgString(svgString: string): SVGElement | null {
@@ -26,6 +33,7 @@ function parseSvgString(svgString: string): SVGElement | null {
 const providerDisplayInfo: Record<string, { name: string; syncType: SyncType }> = {
   "bLyrics-richsynced": { name: "Better Lyrics", syncType: "syllable" },
   "bLyrics-synced": { name: "Better Lyrics", syncType: "line" },
+  "legato-synced": { name: "Better Lyrics Legato", syncType: "line" },
   "musixmatch-richsync": { name: "Musixmatch", syncType: "word" },
   "musixmatch-synced": { name: "Musixmatch", syncType: "line" },
   "lrclib-synced": { name: "LRCLib", syncType: "line" },
@@ -110,6 +118,7 @@ export function addFooter(
     iconWrapper.style.marginLeft = "6px";
     iconWrapper.style.display = "inline-flex";
     iconWrapper.style.verticalAlign = "middle";
+    iconWrapper.style.color = syncTypeColors[info.syncType];
     const svgIcon = parseSvgString(syncTypeIcons[info.syncType]);
     if (svgIcon) {
       iconWrapper.appendChild(svgIcon);
