+++ src/options/editor/features/storage.ts
@@ -3,14 +3,7 @@ import { SYNC_STORAGE_LIMIT, MAX_RETRY_ATTEMPTS, CHUNK_SIZE, LOCAL_STORAGE_SAFE_
 import { syncIndicator } from "../ui/dom";
 import { editorStateManager } from "../core/state";
 import { setThemeName } from "./themes";
-
-async function getLocalStorage<T>(keys: string | string[] | null): Promise<T> {
-  return (await chrome.storage.local.get(keys as string[])) as unknown as T;
-}
-
-async function getSyncStorage<T>(keys: string | string[] | null): Promise<T> {
-  return (await chrome.storage.sync.get(keys as string[])) as unknown as T;
-}
+import { getLocalStorage, getSyncStorage } from "@core/storage";
 
 interface CSSStorageData {
   cssStorageType?: "sync" | "local" | "chunked";
